requirements.txt
uvicorn==0.35.0
fastapi==0.116.1
mcp==1.13.1
pydantic==2.11.7
langgraph==0.6.7
python-dotenv==1.1.1
livekit==1.0.12
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google.generativeai==0.8.5
python-dateutil
sentence-transformers
langmem